Move the node bindings shim out of binary-manager
Ten modules and the test suite imported the injectable fs, path, and child_process bindings from binary-manager.ts. They now live in src/node.ts, so the binary manager can be replaced without dragging every module that only wanted the shim.
