+++ tests/vault-registry.test.ts
@@ -1,5 +1,5 @@
 import { vi, describe, it, expect, beforeEach } from "vitest";
-import { node } from "../src/binary-manager";
+import { node } from "../src/node";
 import {
     VaultRegistry,
     computeVaultId,
